refactor(tree): converge advisor into canonical slice boundary
