+++ calculogic-validator/src/tree-structure-advisor.host.mjs
@@ -1,9 +1,1 @@
-import {
-  runTreeStructureAdvisor as runTreeStructureAdvisorRuntime,
-  summarizeFindings,
-} from './tree-structure-advisor.logic.mjs';
-
-export const runTreeStructureAdvisor = (repositoryRoot, { scope, config, targets } = {}) =>
-  runTreeStructureAdvisorRuntime(repositoryRoot, { scope, config, targets });
-
-export { summarizeFindings };
+export * from './tree/tree-structure-advisor.host.mjs';
