+++ calculogic-validator/src/tree/tree-structure-advisor.contracts.mjs
@@ -1,0 +1,1 @@
+export const TREE_STRUCTURE_ADVISOR_VALIDATOR_ID = 'tree-structure-advisor';
